Remove placeholder Prismic queries on Home page and layout component.
These were initially implemented in order to test out the Prismic
integration with Gatsby and this site. Since we want to initially launch
with just the blog integrated with Prismic, we are removing all the
placeholder queries from the initial development period.

The code being reverted here may still be a useful reference in the
future when we fully integrate the site with Prismic.