Change SF Families logo to their new Our 415 logo.
SF Families rebranded as Our 415, and this is their updated logo.

This also makes a few small tweaks to the PartnersAndSponsorsBlock so
that image heights are capped and properly centered. We missed that
styling previously because all of our previous images were the exact
same dimensions (they had already been resized in Figma), but our
component should handle images of any size without disturbing the
overall layout.