Revert package-lock.json to lockfileVersion 1.
Lockfile version 2 is making it more difficult to upgrade packages since
it is tied to npm versions 7 and 8, but this app still uses npm 6 via
Node.js 14. Currently, upgrading packages is challenging because minor
changes to the requested dependencies cause extremely large cascading
changes to dependent packages, and it is hard to pinpoint exactly which
of the hundreds of upgraded packages is actually causing issues to
surface.

Reverting back to lockfile version 1 allows us to more easily detect
exactly what changes when a dependency is upgrade and will hopefully
facilite upgrading everything, including Node.js, to their latest
versions.