Override gatsby-source-prismic's dependency on gatsby-source-filesystem to a lower version.
gatsby-source-prismic 5.3.1 contains an important fix
(https://github.com/prismicio/prismic-gatsby/pull/532) related to
`@imgix/gatsby`, but contains another commit that unnecessarily depends
on a newer-than-necessary version of gatsby-source-filesystem due to
https://github.com/prismicio/prismic-gatsby/pull/529/files#diff-bf74e96df5fcca670d975bab0f3ff6a636e701d5394a7632b17eebdd05f330e8R54.
This version of gatsby-source-prismic is supposed to still support
Gatsby v4 and gatsby-source-filesystem v4, but due to how it declared
its dependencies, npm will try to use gatsby-source-filesystem v5, which
results in multiple versions of that dependency being used in our Gatsby
v4 app.

This adds an npm override to force gatsby-source-prismic to use the same
version of gatsby-source-filesystem as gatsby itself, which will avoid a
loud warning message when building the site.