Upgrade to storybook 7.4.1.
This was mostly turnkey, since Storybook provides a command for
automatically upgrading and even updating configuration files.

One manual change that was necessary was to update the Storybook webpack
configuration, which was described in terms of patching the existing
webpack configuration. The patched rule changed from array index 0 to 2,
so we had to update the indexes in our webpack configuration. This
matches the changes described in
https://github.com/storybookjs/storybook/issues/22065#issuecomment-1646273441