Update new site CSS to compensate for removal of legacy CSS.
Although we had tried to isolate the legacy CSS styles, some were global
styles that affected the whole site. This updates the remaining CSS
styles in order to retain the look and feel even with the legacy CSS
removed.

One style change that does not match the old site is the link styling.
This actually makes some changes to match the styling shown in the
original Figma designs, but that we were enable to make due to the reset
styles of the previous site being too difficult to override. Now that
they are gone, we can actually implement the styling suggested by the
designs.