Migrate off react-helmet and onto Gatsby Head.
gatsby-plugin-react-helmet is deprecated, and the version of
react-helmet we were using was causing some minor issues with upgrading
the version of React. Since Gatsby is deprecating its official support
for react-helmet, it may be easier to just migrate to Gatsby's new
builtin Head API, which provides similar functionality.

This follows the instructions in the official documentation[^1] for
using the Head API. The main difference from react-helmet is that
instead of defining a base `<Helmet>` tag that can be overwritten on
individual pages through nested `<Helmet>` tags, the Head API requires
exporting a separate `<Head>` component on each page. Although there is
no built-in overrides system, since it just uses plain React components
under the hood, we can build our own, simple system.

We define a `BaseHead` component with all the common tags, and we also
have a mandatory `title` prop, which serves as a reminder for each page
to explicitly define its title. Almost every page today only overrides
the `title`, so this is mostly sufficient. The one exception is the Blog
Post template, which is used to dynamically generate pages based on the
results of GraphQL API queries. On that page, in addition to
instantiating the `BaseHead` component, we add a number of sibling
elements, using page data that Gatsby passes into the `Head` component.

This also adds in the missing title for the Programs page, which was
missing under the react-helmet implementation.

[^1]: https://www.gatsbyjs.com/docs/reference/built-in-components/gatsby-head/