Upgrade to Gatsby 5.
This upgrades to Gatsby 5, bumping all other dependencies that are tied
to the Gatsby version. Most notably, this bumps the version of the
Gatsby-Prismic plugins from 5 to 6, which included a number of breaking
changes:

https://github.com/prismicio/prismic-gatsby/blob/c8e7bbaec5fe4714227d23b0815aabdcec5b2d8d/docs/migration-guide-v5-to-v6.md

All the breaking changes were simple to resolve (even if debugging them
may have been challenging), and the site and the Prismic previews
functionality still work exactly as they did before.