Add `@csstools/postcss-global-data` dependency.
The recent update to PostCSS and postcss-custom-media appears to have
broken the latter, causing custom CSS media queries to not be processed.
It's not clear exactly what changed that caused it, but it may have been
one of nested packages.

Regardless, it looks like the postcss-custom-media project's README was
updated sometime earlier this year to explicitly mention that it may not
work out of the box with CSS Modules, which we are using in this
project. They recommend using the postcss-global-data package to ensure
that the custom media definitions are available in all CSS modules.

Adding that package appears to fix the issue, allowing the pages to look
the same as before the package upgrades.